Remove unrelated code in typed-store
