+++ Cargo.toml
@@ -2,7 +2,6 @@
 default-members = [
   "crates/checkpoint-downloader",
   "crates/typed-store",
-  "crates/typed-store-derive",
   "crates/walrus-core",
   "crates/walrus-e2e-tests",
   "crates/walrus-proc-macros",
@@ -134,7 +133,6 @@ tracing-opentelemetry = { version = "=0.28.0", default-features = false }
 tracing-subscriber = { version = "0.3.19", default-features = false, features = ["fmt"] }
 twox-hash = "2.1.0"
 typed-store = { path = "crates/typed-store" }
-typed-store-derive = { path = "crates/typed-store-derive" }
 uint = "0.10.0"
 # TODO (WAL-790): update typed-store to >= 1.47 or remove it.
 url = "2.5.4"
