--- crates/typed-store-derive/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "typed-store-derive"
-publish = false
-authors.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-itertools.workspace = true
-proc-macro2.workspace = true
-quote.workspace = true
-syn = { workspace = true, features = ["full"] }
-
-
-[lints]
-workspace = true
