+++ crates/typed-store/src/rocks/safe_iter.rs
@@ -1,4 +1,4 @@
-// Copyright (c) Mysten Labs, Inc.
+// Copyright (c) Walrus Foundation
 // SPDX-License-Identifier: Apache-2.0
 
 use std::{fmt, marker::PhantomData, sync::Arc};
