+++ crates/typed-store/src/traits.rs
@@ -1,7 +1,7 @@
-// Copyright (c) Mysten Labs, Inc.
+// Copyright (c) Walrus Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{borrow::Borrow, collections::BTreeMap, error::Error, ops::RangeBounds};
+use std::{borrow::Borrow, error::Error, ops::RangeBounds};
 
 use serde::{Serialize, de::DeserializeOwned};
 
@@ -113,27 +113,4 @@ pub struct TableSummary {
     pub key_hist: hdrhistogram::Histogram<u64>,
     /// The histogram of value sizes
     pub value_hist: hdrhistogram::Histogram<u64>,
-}
-
-/// The trait for the typed store to dump the table
-pub trait TypedStoreDebug {
-    /// Dump a DB table with pagination
-    fn dump_table(
-        &self,
-        table_name: String,
-        page_size: u16,
-        page_number: usize,
-    ) -> eyre::Result<BTreeMap<String, String>>;
-
-    /// Get the name of the DB. This is simply the name of the struct
-    fn primary_db_name(&self) -> String;
-
-    /// Get a map of table names to key-value types
-    fn describe_all_tables(&self) -> BTreeMap<String, (String, String)>;
-
-    /// Count the entries in the table
-    fn count_table_keys(&self, table_name: String) -> eyre::Result<usize>;
-
-    /// Return table summary of the input table
-    fn table_summary(&self, table_name: String) -> eyre::Result<TableSummary>;
 }
